requirements.txt
openai
faiss-cpu
streamlit
pymupdf
requests
python-dotenv
tiktoken
